Add Nanoc::Spec
